+++ calyx-opt/src/passes/compile_static_interface.rs
@@ -1,12 +1,9 @@
 use crate::analysis::StaticSchedule;
-use crate::passes::math_utilities::get_bit_width_from;
 use crate::traversal::{Action, Named, VisResult, Visitor};
 use calyx_ir as ir;
 use ir::{
-    build_assignments, guard, structure, Attributes, Guard, Nothing,
-    StaticTiming, RRC,
+    build_assignments, guard, structure, Attributes, Nothing, StaticTiming, RRC,
 };
-use itertools::Itertools;
 use std::cell::RefCell;
 use std::rc::Rc;
 
@@ -20,52 +17,6 @@ impl Named for CompileStaticInterface {
 
     fn description() -> &'static str {
         "Compiles Static Component Interface"
-    }
-}
-
-// Looks recursively thru guard to %[0:n] into %0 | %[1:n].
-fn separate_first_cycle(
-    guard: ir::Guard<StaticTiming>,
-) -> ir::Guard<StaticTiming> {
-    match guard {
-        ir::Guard::Info(st) => {
-            let (beg, end) = st.get_interval();
-            if beg == 0 && end != 1 {
-                let first_cycle =
-                    ir::Guard::Info(ir::StaticTiming::new((0, 1)));
-                let after = ir::Guard::Info(ir::StaticTiming::new((1, end)));
-                let cong = ir::Guard::or(first_cycle, after);
-                return cong;
-            }
-            guard
-        }
-        ir::Guard::And(l, r) => {
-            let left = separate_first_cycle(*l);
-            let right = separate_first_cycle(*r);
-            ir::Guard::and(left, right)
-        }
-        ir::Guard::Or(l, r) => {
-            let left = separate_first_cycle(*l);
-            let right = separate_first_cycle(*r);
-            ir::Guard::or(left, right)
-        }
-        ir::Guard::Not(g) => {
-            let a = separate_first_cycle(*g);
-            ir::Guard::Not(Box::new(a))
-        }
-        _ => guard,
-    }
-}
-
-// Looks recursively thru assignment's guard to %[0:n] into %0 | %[1:n].
-fn separate_first_cycle_assign(
-    assign: ir::Assignment<StaticTiming>,
-) -> ir::Assignment<StaticTiming> {
-    ir::Assignment {
-        src: assign.src,
-        dst: assign.dst,
-        attributes: assign.attributes,
-        guard: Box::new(separate_first_cycle(*assign.guard)),
     }
 }
 
@@ -126,80 +77,6 @@ fn make_assign_dyn_one_cycle_static_comp(
 }
 
 impl CompileStaticInterface {
-    // Takes the assignments within a static component, and instantiates
-    // an FSM (i.e., counter) to convert %[i:j] into i<= fsm < j.
-    // Also includes logic to make fsm reset to 0 once it gets to n-1.
-    fn make_early_reset_assigns_static_component(
-        &mut self,
-        sgroup_assigns: &mut Vec<ir::Assignment<ir::StaticTiming>>,
-        latency: u64,
-        fsm: ir::RRC<ir::Cell>,
-        builder: &mut ir::Builder,
-        comp_sig: RRC<ir::Cell>,
-    ) -> Vec<ir::Assignment<Nothing>> {
-        let fsm_name = fsm.borrow().name();
-        let fsm_size = fsm
-            .borrow()
-            .find("out")
-            .unwrap_or_else(|| unreachable!("no `out` port on {fsm_name}"))
-            .borrow()
-            .width;
-        let mut assigns = sgroup_assigns
-            .drain(..)
-            .map(separate_first_cycle_assign)
-            .collect_vec();
-        let mut dyn_assigns = assigns
-            .drain(..)
-            .map(|assign| {
-                StaticSchedule::make_assign_dyn(
-                    assign,
-                    &fsm,
-                    fsm_size,
-                    builder,
-                    true,
-                    Some(Rc::clone(&comp_sig)),
-                )
-            })
-            .collect_vec();
-        let this = Rc::clone(&comp_sig);
-        structure!( builder;
-            // done hole will be undefined bc of early reset
-            let signal_on = constant(1,1);
-            let adder = prim std_add(fsm_size);
-            let const_one = constant(1, fsm_size);
-            let first_state = constant(0, fsm_size);
-            let final_state = constant(latency-1, fsm_size);
-        );
-        let g1: Guard<Nothing> = guard!(this["go"]);
-        let g2: Guard<Nothing> = guard!(fsm["out"] == first_state["out"]);
-        let trigger_guard = ir::Guard::and(g1, g2);
-        let g3: Guard<Nothing> = guard!(fsm["out"] != first_state["out"]);
-        let g4: Guard<Nothing> = guard!(fsm["out"] != final_state["out"]);
-        let incr_guard = ir::Guard::and(g3, g4);
-        let stop_guard: Guard<Nothing> =
-            guard!(fsm["out"] == final_state["out"]);
-        let fsm_incr_assigns = build_assignments!(
-          builder;
-          // Incrementsthe fsm
-          adder["left"] = ? fsm["out"];
-          adder["right"] = ? const_one["out"];
-          // Always write into fsm.
-          fsm["write_en"] = ? signal_on["out"];
-          // If fsm == 0 and comp.go is high, then we start an execution.
-          fsm["in"] = trigger_guard ? const_one["out"];
-          // If 1 < fsm < n - 1, then we unconditionally increment the fsm.
-          fsm["in"] = incr_guard ? adder["out"];
-          // If fsm == n -1 , then we reset the FSM.
-          fsm["in"] = stop_guard ? first_state["out"];
-          // Otherwise the FSM is not assigned to, so it defaults to 0.
-          // If we want, we could add an explicit assignment here that sets it
-          // to zero.
-        );
-        dyn_assigns.extend(fsm_incr_assigns);
-
-        dyn_assigns
-    }
-
     // Makes `done` signal for promoted static<n> component.
     fn make_done_signal_for_promoted_component(
         &mut self,
@@ -278,8 +155,6 @@ impl Visitor for CompileStaticInterface {
         _comps: &[ir::Component],
     ) -> VisResult {
         if comp.is_static() && s.get_latency() > 1 {
-            // Handle components with latency > 1.
-            let latency = s.get_latency();
             if let ir::StaticControl::Enable(sen) = s {
                 let mut sch = StaticSchedule::from(vec![Rc::clone(&sen.group)]);
                 let mut builder = ir::Builder::new(comp, sigs);
